+++ src/main/java/com/synopsys/integration/detect/tool/signaturescanner/enums/SignatureScanStatusType.java
@@ -1,0 +1,7 @@
+package com.synopsys.integration.detect.tool.signaturescanner.enums;
+
+public enum SignatureScanStatusType {
+    SUCCESS,
+    FAILURE,
+    SKIPPED
+}