+++ src/main/java/com/synopsys/integration/detect/workflow/blackduck/codelocation/WaitableCodeLocationData.java
@@ -4,12 +4,12 @@ import java.util.Set;
 
 import com.synopsys.integration.blackduck.service.model.NotificationTaskRange;
 
-public class AccumulatedCodeLocationData {
+public class WaitableCodeLocationData {
     private final int expectedNotificationCount;
     private final Set<String> successfulCodeLocationNames;
     private final NotificationTaskRange notificationTaskRange;
 
-    public AccumulatedCodeLocationData(int expectedNotificationCount, Set<String> successfulCodeLocationNames, NotificationTaskRange notificationTaskRange) {
+    public WaitableCodeLocationData(int expectedNotificationCount, Set<String> successfulCodeLocationNames, NotificationTaskRange notificationTaskRange) {
         this.expectedNotificationCount = expectedNotificationCount;
         this.successfulCodeLocationNames = successfulCodeLocationNames;
         this.notificationTaskRange = notificationTaskRange;